refactor: 할당중인 차량 리스트 조회 기능을 reservation에서 vehicle로 이동
