+++ domain/src/main/java/com/pickcar/reservation/infrastructure/ReservationRepository.java
@@ -41,12 +41,6 @@ public interface ReservationRepository extends JpaRepository<Reservation, Long> 
     List<EmployeeReservationProjection> findEmployeesWithReservationPreInfo(UserRole role,
                                                                             List<ReservationStatus> statuses);
 
-    @Query("SELECT v FROM Vehicle v " +
-            "WHERE v.status = :vehicleStatus " +
-            "AND EXISTS (SELECT r FROM Reservation r " +
-            "WHERE r.vehicleId = v.id AND r.status IN :reservationStatuses)")
-    List<Vehicle> findAssignedVehicles(VehicleStatus vehicleStatus, List<ReservationStatus> reservationStatuses);
-
     @Query("SELECT r.id FROM Reservation r WHERE r.vehicleId = :vehicleId " +
             "AND r.userId = :userId AND r.status IN :statuses")
     Optional<Long> findIdByVehicleIdAndUserIdAndStatusIn(Long vehicleId, Long userId,
