+++ domain/src/main/java/com/pickcar/vehicle/application/VehicleService.java
@@ -1,5 +1,6 @@
 package com.pickcar.vehicle.application;
 
+import com.pickcar.reservation.domain.ReservationStatus;
 import com.pickcar.vehicle.domain.Vehicle;
 import com.pickcar.vehicle.domain.VehicleInfo;
 import com.pickcar.vehicle.domain.VehicleStatus;
@@ -8,6 +9,7 @@ import com.pickcar.vehicle.exception.VehicleException;
 import com.pickcar.vehicle.infrastructure.VehicleRepository;
 import com.pickcar.vehicle.presentation.dto.request.ChangeVehicleStatusRequest;
 import com.pickcar.vehicle.presentation.dto.request.VehicleRegisterRequest;
+import com.pickcar.vehicle.presentation.dto.response.SearchAbleVehiclesResponse;
 import com.pickcar.vehicle.presentation.dto.response.UnAllocatedVehicleResponse;
 import com.pickcar.vehicle.presentation.dto.response.VehicleListResponse;
 import java.time.LocalDate;
@@ -106,5 +108,15 @@ public class VehicleService {
         if (!vehicle.tryMarkAsReturned()) {
             throw new VehicleException(VehicleErrorCode.ALREADY_RENTED_OR_RETURNED);
         }
+    }
+
+    public List<SearchAbleVehiclesResponse> getAssignedVehicles() {
+        //운행 가능한 상태의 차면서 예약 상태인 것
+        List<Vehicle> availableVehicles = vehicleRepository.findAssignedVehicles(VehicleStatus.OPERABLE,
+                List.of(ReservationStatus.RESERVED, ReservationStatus.DELAYED));
+
+        return availableVehicles.stream()
+                .map(SearchAbleVehiclesResponse::from)
+                .toList();
     }
 }
